backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Redis and Celery
redis==5.0.1
celery==5.3.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client
httpx==0.25.2

# Validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.7

# Performance monitoring
psutil==5.9.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0